+++ example/jose_example.dart
@@ -277,7 +277,7 @@ void example7() async {
   builder.jsonContent = claims.toJson();
 
   // add a key to sign, can only add one for JWT
-  var key = _readPrivateKeyFromFile('example/jwtRS512.key');
+  var key = JsonWebKey.fromPem(File('example/jwtRS512.key').readAsStringSync());
   builder.addRecipient(key, algorithm: 'RS512');
 
   // build the jws
@@ -300,35 +300,4 @@ void example8() async {
   final valid = key.verify(hash, sig);
 
   print('valid? $valid');
-}
-
-JsonWebKey _readPrivateKeyFromFile(String path) {
-  var v = parsePem(File(path).readAsStringSync()).first;
-  var keyPair = (v is PrivateKeyInfo) ? v.keyPair : v as KeyPair;
-  var pKey = keyPair.privateKey as RsaPrivateKey;
-  print(pKey);
-
-  String _bytesToBase64(List<int> bytes) {
-    return base64Url.encode(bytes).replaceAll('=', '');
-  }
-
-  String _intToBase64(BigInt v) {
-    return _bytesToBase64(v
-        .toRadixString(16)
-        .replaceAllMapped(RegExp('[0-9a-f]{2}'), (m) => '${m.group(0)},')
-        .split(',')
-        .where((v) => v.isNotEmpty)
-        .map((v) => int.parse(v, radix: 16))
-        .toList());
-  }
-
-  return JsonWebKey.fromJson({
-    'kty': 'RSA',
-    'n': _intToBase64(pKey.modulus),
-    'd': _intToBase64(pKey.privateExponent),
-    'p': _intToBase64(pKey.firstPrimeFactor),
-    'q': _intToBase64(pKey.secondPrimeFactor),
-    'alg': 'RS512',
-    'kid': 'some_id'
-  });
 }